chirpchat-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
cloudinary==1.36.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
sqlalchemy[asyncio]
asyncpg
alembic
psycopg2-binary
fastapi-mail==1.4.1
pywebpush
pytz==2024.1
redis[hiredis]==5.0.1
sse-starlette==2.1.0
prometheus-fastapi-instrumentator
mutagen
aiofiles
python-magic-bin
firebase-admin==6.4.0

    